move polygon to 4-2 package
